+++ R/RcppExports.R
@@ -172,6 +172,10 @@ vaeInnerSetup_ <- function(e) {
     .Call(`_nlmixr2est_vaeInnerSetup_`, e)
 }
 
+vaeInnerUpdatePar_ <- function(thFull, omegaDiag) {
+    .Call(`_nlmixr2est_vaeInnerUpdatePar_`, thFull, omegaDiag)
+}
+
 vaeInnerLik <- function(etaMat, cores, grad = FALSE, preds = FALSE) {
     .Call(`_nlmixr2est_vaeInnerLik`, etaMat, cores, grad, preds)
 }
@@ -206,6 +210,10 @@ vaeIterPrintRow_ <- function(x, f, phase = "") {
 
 vaeIterPrintGet_ <- function(printLine = TRUE) {
     .Call(`_nlmixr2est_vaeIterPrintGet_`, printLine)
+}
+
+vaeTrainCpp_ <- function(params, prep, control, nMix, mixProbR, cores, row0, parNames, iterPrintControl, xform, structIdx0) {
+    .Call(`_nlmixr2est_vaeTrainCpp_`, params, prep, control, nMix, mixProbR, cores, row0, parNames, iterPrintControl, xform, structIdx0)
 }
 
 boxCox_ <- function(x = 1L, lambda = 1, yj = 0L) {